+++ lit_gpt/datasets/__init__.py
@@ -1,4 +1,4 @@
 # Copyright Lightning AI. Licensed under the Apache License 2.0, see LICENSE file.
 
-from lit_gpt.datasets.base import LitDataModule
+from lit_gpt.datasets.data_module import LitDataModule
 from lit_gpt.datasets.tinyllama import TinyLlama
